+++ src/main/java/de/maxhenkel/enhancedgroups/EnhancedGroups.java
@@ -1,14 +1,13 @@
 package de.maxhenkel.enhancedgroups;
 
+import de.maxhenkel.admiral.MinecraftAdmiral;
 import de.maxhenkel.configbuilder.ConfigBuilder;
-import de.maxhenkel.enhancedgroups.command.AutoJoinGroupCommands;
-import de.maxhenkel.enhancedgroups.command.ForceJoinCommands;
-import de.maxhenkel.enhancedgroups.command.InstantGroupCommands;
-import de.maxhenkel.enhancedgroups.command.PersistentGroupCommands;
+import de.maxhenkel.enhancedgroups.command.*;
 import de.maxhenkel.enhancedgroups.config.AutoJoinGroupStore;
 import de.maxhenkel.enhancedgroups.config.CommonConfig;
 import de.maxhenkel.enhancedgroups.config.PersistentGroupStore;
 import de.maxhenkel.enhancedgroups.events.GroupSummaryEvents;
+import de.maxhenkel.voicechat.api.Group;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import org.slf4j.Logger;
@@ -26,17 +25,27 @@ public class EnhancedGroups implements ModInitializer {
     public static PersistentGroupStore PERSISTENT_GROUP_STORE;
     public static AutoJoinGroupStore AUTO_JOIN_GROUP_STORE;
 
+    public static EnhancedGroupPermissionManager PERMISSION_MANAGER;
+
     @Override
     public void onInitialize() {
         Path configFolder = Paths.get(".", "config").resolve(MOD_ID);
         CONFIG = ConfigBuilder.build(configFolder.resolve("%s.properties".formatted(MOD_ID)), true, CommonConfig::new);
         PERSISTENT_GROUP_STORE = new PersistentGroupStore(configFolder.resolve("persistent-groups.json").toFile());
         AUTO_JOIN_GROUP_STORE = new AutoJoinGroupStore(configFolder.resolve("auto-join-groups.json").toFile());
+        PERMISSION_MANAGER = new EnhancedGroupPermissionManager();
+
         CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> {
-            InstantGroupCommands.register(dispatcher);
-            PersistentGroupCommands.register(dispatcher);
-            AutoJoinGroupCommands.register(dispatcher);
-            ForceJoinCommands.register(dispatcher);
+            MinecraftAdmiral.builder(dispatcher, registryAccess)
+                    .addCommandClasses(
+                            AutoJoinGroupCommands.class,
+                            ForceJoinCommands.class,
+                            InstantGroupCommands.class,
+                            PersistentGroupCommands.class
+                    )
+                    .setPermissionManager(PERMISSION_MANAGER)
+                    .addArgumentTypes(argumentTypeRegistry -> argumentTypeRegistry.register(Group.Type.class, new GroupTypeArgumentSupplier(), new GroupTypeArgumentTypeSupplier()))
+                    .build();
         });
         GroupSummaryEvents.init();
     }